Move more Cassandra code out of Data
