+++ services/galley/src/Galley/API/Create.hs
@@ -39,7 +39,6 @@ import Galley.API.Mapping
 import Galley.API.One2One
 import Galley.API.Util
 import Galley.App
-import Galley.Data.Access
 import qualified Galley.Data.Conversation as Data
 import Galley.Data.Conversation.Types
 import Galley.Effects
@@ -530,11 +529,11 @@ toUUIDs a b = do
   return (a', b')
 
 accessRole :: NewConv -> AccessRole
-accessRole b = fromMaybe defRole (newConvAccessRole b)
+accessRole b = fromMaybe Data.defRole (newConvAccessRole b)
 
 access :: NewConv -> [Access]
 access a = case Set.toList (newConvAccess a) of
-  [] -> defRegularConvAccess
+  [] -> Data.defRegularConvAccess
   (x : xs) -> x : xs
 
 newConvMembers :: Local x -> NewConv -> UserList UserId
